rename invokeShortLink and move it to GraphQL
