+++ frontend/src/service/shortGraphQL/ShortLinkGraphQL.api.ts
@@ -4,11 +4,12 @@ import { GraphQLService, IGraphQLRequestError } from '../GraphQL.service';
 import { ShortLink } from '../../entity/ShortLink';
 import { getErrorCodes } from '../GraphQLError';
 import {
+  IShortGraphQLMutation,
   IShortGraphQLQuery,
   IShortGraphQLShortLink,
   IShortGraphQLShortLinkInput
 } from './schema';
-import { CaptchaService, UPDATE_SHORT_LINK } from '../Captcha.service';
+import {CaptchaService, CREATE_SHORT_LINK, UPDATE_SHORT_LINK} from '../Captcha.service';
 
 export class ShortLinkGraphQLApi {
   private readonly baseURL: string;
@@ -49,6 +50,89 @@ export class ShortLinkGraphQLApi {
           reject(errCodes[0]);
         });
     });
+  }
+
+  async createShortLink(
+      shortLink: ShortLink,
+      isPublic: boolean
+  ): Promise<ShortLink> {
+    const gqlCreateShortLink = `
+      mutation params(
+        $captchaResponse: String!
+        $authToken: String!
+        $shortLinkInput: ShortLinkInput!
+        $isPublic: Boolean!
+      ) {
+        authMutation(authToken: $authToken, captchaResponse: $captchaResponse) {
+          createShortLink(shortLink: $shortLinkInput, isPublic: $isPublic) {
+            alias
+            longLink
+          }
+        }
+      }
+    `;
+
+    let captchaResponse = '';
+
+    try {
+      captchaResponse = await this.captchaService.execute(CREATE_SHORT_LINK);
+    } catch (err) {
+      return Promise.reject(err);
+    }
+
+    const variables = this.gqlCreateShortLinkVariable(
+        captchaResponse,
+        shortLink,
+        isPublic
+    );
+    return new Promise<ShortLink>(
+        (
+            resolve: (createdShortLink: ShortLink) => void,
+            // TODO(task#h6V56gf9): change the string type to Err for this function and all callers
+            reject: (errCode: string) => any
+        ) => {
+          this.graphQLService
+              .mutate<IShortGraphQLMutation>(this.baseURL, {
+                mutation: gqlCreateShortLink,
+                variables: variables
+              })
+              .then((res: IShortGraphQLMutation) => {
+                const shortLink = this.getShortLinkFromCreatedShortLink(
+                    res.authMutation.createShortLink
+                );
+                resolve(shortLink);
+              })
+              .catch((err: IGraphQLRequestError) => {
+                const errCodes = getErrorCodes(err);
+                reject(errCodes[0]);
+              });
+        }
+    );
+  }
+
+  private getShortLinkFromCreatedShortLink(
+      createdShortLink: IShortGraphQLShortLink
+  ): ShortLink {
+    return {
+      originalUrl: createdShortLink.longLink,
+      alias: createdShortLink.alias
+    };
+  }
+
+  private gqlCreateShortLinkVariable(
+      captchaResponse: string,
+      link: ShortLink,
+      isPublic: boolean = false
+  ) {
+    return {
+      captchaResponse: captchaResponse,
+      authToken: this.authService.getAuthToken(),
+      shortLinkInput: {
+        longLink: link.originalUrl,
+        customAlias: link.alias
+      },
+      isPublic
+    };
   }
 
   async updateShortLink(oldAlias: string, shortLink: Partial<ShortLink>) {
